+++ apps/transcoder/src/main/java/com/ott/transcoder/ffmpeg/Resolution.java
@@ -1,4 +1,4 @@
-package com.ott.domain.video_profile.domain;
+package com.ott.transcoder.ffmpeg;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
